pixela-java-client-63 Refactoring Codec classes - move json interfaces to pixela.client.http.json package.
