+++ default-impl/src/test/java/pixela/client/impl/JdkPostRequestBuilderTest.java
@@ -23,6 +23,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import pixela.client.UserToken;
 import pixela.client.http.Post;
 import pixela.client.http.Request;
+import pixela.client.http.json.JsonEncoder;
 import reactor.adapter.JdkFlowAdapter;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
